♻️ refactor : 업로드 로직 변경 #275
