+++ src/hooks/use-slugged-save-card.ts
@@ -8,10 +8,10 @@ import { checkSlugExists } from '@/apis/check-slug-exists';
 import { useStageRefStore } from '@/store/editor.stage.store';
 import { ROUTES } from '@/constants/path.constant';
 import { useRouter } from 'next/navigation';
-import { validateSlug } from '@/utils/editor/validate-slug';
+import { validateSlug } from '@/utils/editor/save/validate-slug';
 import { resetEditorState } from '@/utils/editor/editor-reset-state';
-import { createCardInsertPayload } from '@/utils/editor/create-card-insert-payload';
-import { getStageImageUrls } from '@/utils/editor/get-State-Image-urls';
+import { createCardInsertPayload } from '@/utils/editor/save/create-card-insert-payload';
+import { getStageImageUrls } from '@/utils/editor/save/get-State-Image-urls';
 
 export const useSluggedSaveCard = () => {
   const router = useRouter();
